Level 3 constructed, level 2 bug
spikes don't have functionality yet, but objects have been placed and have the mask renderer functionality.
Objects Camera does not work as intended- it moves to objects as intended with a mouse click, but moves the green object with it. Green object should not move with camera. Everything else is functional, but level 3 is not yet implemented in ProgressLevel() function (Level 3 currently loads win screen, not the third area). Player should transport to level 3 when level 2 is complete.

Win condition for level 3 not implemented. Doorway, UI for password, and numbers/ letters on objects aren't there yet.